Revert "VM: Introduce writePayloadRpData/readPayloadRpData"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5519e63d-81ff-4c41-a830-f8157b11b390). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0a4c19eedf278c2da3eb2c35b5552204b2114500
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com