Benchmark rollbackProtectedSecret API
This gives the rough estimate of read/write latency. Note that the
implementation of the API includes refresh in case of expired session if
that happens Secretkeeper is called twice (the first call fails session
is refreshed).

In order to ensure a good mix of retried request, Vms are started
asynchronously, the threads wait for random time before starting the
payload.

Test: MicrodroidBenchmarks#testPayloadRpDataLatency
Bug: 379073746
Change-Id: Ia05e9bc0b0b13b2e90fa308295966e592669e94b