Set min_sdk_version for MicrodroidTestNativeLib
MicrodroidTestNativeLib is used as a jni lib for MicrodroidTestApp
that sets min_sdk_version to "33". However, missing min_sdk_version
of MicrodroidTestNativeLib leads to link with strong symbols that are
not compatible with the old Android versions.
Adding the same min_sdk_version makes the symbols weak and compatible
with the old Android versions.

Bug: 388693148
Test: check build command includes correct clang triple.
Change-Id: I4dab64fb812e08b27557574a48266daf298e3f14