dice: add sign and encode with CoseSign1
Adds APIs to produce CoseSign1 encoded signatures given a message,
aad, and private key. This is implemented with
`DiceCoseSignAndEncodeSign1` from open_dice, which is brought in through
bindgen build rules for soong and trusty.

The driving use case is pvmdice, which needs to implement the hwbcc
interface's sign_data function. This will enable RKP protocols as well
as the implementation of the authmgr frontend in trusty pVMs.

Bug: 390385820
Test: atest libdiced_open_dice.integration_test
Test: atest libdiced_open_dice_nostd.integration_test
Test: build.py qemu-generic-arm64-test-debug \
        --test="boot-test:com.android.trusty.rust.diced_open_dice_tests.test"

Change-Id: Iac77f2f48f5f9ae90d1d08aaa6f34298b7220e40