Revert "Remove unnecessary @JvmStatic"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/abfa212e-3e71-4e3f-92d5-8f47e8983b65). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iea6a21ed4e0af6dea63c126cc557fd112df1bffa
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com