Revert "Avoid using static mut variables."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/82edd429-ab38-40b2-bac3-4df0e4b8fd6a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If8f45790393c64f2f9a1e82dd438820757e40026
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com