Revert "Update kernel to builds 12934375"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f57cbc15-22a4-4c02-842c-d6adc1ffa2cf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 391036599
Change-Id: Ib52eac2a6b87f519413885e7032cf241c38a6bc0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com