Revert "Update kernel to builds 12934375"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b2681777-92f7-49b5-b79d-d9abab047c1a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 391169593
Change-Id: Ib49c3e4e2f2f57c7cd9eb6a7222ed71b2286be4a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com