Revert "Convert TerminalView to kotlin"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1a779700-8488-4645-b3fd-5c6a4e81773b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I48c01bd5b1d3b191fff218cea9b00715be84a6b5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com