Revert "dice: add trusty build rules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/20bc2b14-6f93-4aa2-a826-193a5b25adfa). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I718701760a85392dca658c60ca6dda2a52219238
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com