Revert "dice: add trusty build rules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e36b8249-cb45-48f0-9174-221eb0cb0811). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ida4896f59333916d0426743f8739e9e8effb4dad
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com