Revert "libvmclient: Take VmCallback in VmInstance::start()"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/21ed2dd7-4276-4727-9e3e-639631355f3b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 391356639
Change-Id: I3c2d32be6dedc14ec56732a81abcf8d700fd27e5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com