Revert "Avoid using static mut variables."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e798bd3b-5738-4e29-930e-483826a3aa91). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib9e3b2a12fa7f6657f7ebefb3d3aa872ac62d4a7
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com