Revert "Use &raw rather than addr_of macros."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b329b8d7-5e4f-4ca7-aa01-72041e6d1c7e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I427d2e0af7c82dd0f74117ef3a92b41bffa11341
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com