Revert "Avoid using static mut variables."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b329b8d7-5e4f-4ca7-aa01-72041e6d1c7e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I01659122d2072c9ea6e25679dd656cee6ea05a99
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com