Revert "dice: add sign and encode with CoseSign1"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/85b12ca4-991c-401d-b7f0-94d532f3929b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 391631520
Change-Id: I11156840a65258d2683ac52d0142af870836930a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com