Revert "Resolve build warning in Terminal and its test"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a2e9c899-92b5-4a98-bc15-9a1c9cdcc2f8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie926b47f0290357a7e1e37ff039b259050357210
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com