Revert "Enable SVE when available in AVF"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/78d73ac5-03ba-4dda-951d-69383bd7001f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 391967232
Change-Id: I9dbbc225555eea8bb292083cb92616c980a4cfa0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com