Revert "Enable SVE when available in AVF"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/761e5078-daea-4a7d-8170-9c6082509c3a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 391967128
Change-Id: I5420535d3dfd6665c167537f68f0f999c0612ad9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com