Revert "Enable SVE when available in AVF"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/eacdbc6f-ef72-40b6-8645-0f7a52560fcd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 391972387
Change-Id: Ic445a1414c1894468b05e7173b9945f127cc6227
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com