Revert "Install weston and xwayland by default"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/940b9c8d-c250-4a38-bae1-1ba47a8e0496). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0f90e486ee9586a9c23119b7a461707c2c7561ab
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com