Revert "Enable SVE when available in AVF"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/199b5213-90b1-4b2a-bfc4-4061aea55a24). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 391975793
Change-Id: Ie0aef14388db578533531a45f59f5d1b5b42f956
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com