Revert "Enable SVE when available in AVF"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b53a6796-73b3-4e79-87e1-e9e6951a971e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 391974758
Change-Id: I3828a398a898ded3d4bbfa2e4baee986039789c2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com