Revert "Enable SVE when available in AVF"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1ee0160e-f183-4e24-b677-3031409d68c0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 391974476
Change-Id: I7ad4b1eb8c07ede29767627681b0cafbb809becb
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com