Revert "Enable SVE when available in AVF"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/70cb8808-c2a6-47d8-87c6-3b0e3e035b1c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 391974982
Change-Id: I07c0169914158b44ec59a46d9e534b907df67837
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com