Revert "Enable SVE when available in AVF"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6f37d5f1-c485-494e-a757-2af73079590d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 391980842
Change-Id: I88827a31e724fca61cbaca623691a6cb7a882689
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com