Revert "Enable SVE when available in AVF"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a2a73301-d7a1-4913-a2a1-90cb1e306f85). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 391974297
Change-Id: I917bc6a6e71d30cbc92c4d3ccdde1c4322cb7a53
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com