Revert "Enable SVE when available in AVF"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1749ba11-caba-4718-bd7d-7681c46e193c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 391979538
Bug: 391978891
Change-Id: Ic60f0f15f7fd58a37fb9f6d780713a86d5044dde
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com