Revert "Enable SVE when available in AVF"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/58c47d03-f708-4cc9-ab9f-4c3dde875c59). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 391984981
Change-Id: Icf6e99267a6b459cd409ed08697a8b7f5774eefa
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com