Revert "Updates for virtio-drivers 0.8."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/01e3a3d2-9316-445e-9388-8e449ef2f0aa). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 392168474
Change-Id: Ica057678d2a3a938ce5c5a6f58b4f7445b0ed39b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com