Revert "Import translations. DO NOT MERGE ANYWHERE"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6ed2c544-b043-4291-9262-2c87049984b0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I813d2868f5c59fe107aa2d23b04b7ba3709fa85a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com