Revert "Install weston and xwayland by default"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/093df296-0c7e-4cf0-8817-57344d7f84d7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4f3df88ab555c6152182fc3ad1f2183bfa9ee96a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com