pvmfw: rollback: Prepare for fixed-digest RBP
Refactor the RKP VM fixed-index RBP solution to allow the high-level
logic to support hard-coding other criteria such as a fixed digest.

For robustness, add the expected public key to the check for fixed
index. Note that this is currently superfluous given the way we only
pass PUBLIC_KEY to verify_payload() in main() but makes this logic
future-proof and the assumptions for fixed-index RBP clearer.

Note: No functional change intended.

Bug: 392628867
Bug: 377276983
Test: m pvmfw_bin # Counting on TH for RKP tests
Change-Id: I037b0d0cfee5d7a922d5bea202c06e6902e439a8