build/debian: Fix the resolution of build script's directory
The `$(dirname "$0")` pattern used throughout build.sh does not reliably point to the build script's directory. The linked bug report highlights an issue where `$(dirname "$0")` captures a relative path that is invalidated after changing the current working directory.

We fix this by capturing the absolute path to the build script's directory at the beginning of the build script, and using it in place of the existing `$(dirname "$0")` calls.

We also fix the build.sh invocation within build_in_container.sh to mimic our Kokoro build scripts.

Bug: 392644919
Change-Id: Ia1623a16e9869a30e374577bd76802d2339f1b7b
Test: ./build_in_container.sh -k
Test: ./build_in_container.sh -k -a aarch64