Revert "Explicitly set target_sdk_version of vmshareapp"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/edacd3d3-fb84-48bb-b2a3-cf165c1ec4d5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia8a345d633e279392445b70fa132fb77da90530e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com