Revert "Set min_sdk_version for the test payload"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/edacd3d3-fb84-48bb-b2a3-cf165c1ec4d5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifffd94ad276445f7ad199ff98952c9ee62118e3d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com