apkmanifest: parse fields related to relaxed rollback protection scheme
In case a Microdroid pVM wants to opt in a relaxed rollback protection
scheme it needs to have the following things defined in it's manifest:
* <uses-permission USE_RELAXED_MICRODROID_ROLLBACK_PROTECTION>
* set android.system.virtualmachine.ROLLBACK_INDEX <property>

In case only one of the two things is defined, the VM won't boot. This
is enforced by microdroid_manager (see changes to the verify.rs).

In the follow-up patch these new fields will be used to create a new
more relaxed sealing policy.

Bug: 378681279
Test: atest MicrodroidTests
Change-Id: Iabd12fd47f0eb271f021d5ad466de4f6c0669f2b