pvmfw: Handle absence of vcpu-stall-detector to avoid boot failure
In some VM configurations, the `qemu,vcpu-stall-detector` may not be
used. This change ensures that the absence of a compatible vCPU stall
detector does not cause a boot failure.

When no compatible node is found, a warning is logged and the process
continues without attempting to patch the `interrupts` property. This
allows for smoother boot operations in VMs that do not use this
compatibility feature.

Change-Id: I8094f782345cbe8f4d7a13a81ca44aee72764540