Revert "build/debian: Fix the resolution of build script's directory"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3d598791-a383-4bf3-aee7-15f4b146df91). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I62bcc66e37e5347ae63a90d17c33e2b752369b24
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com