Revert "libvm_payload: Use Rust stubs instead of cc"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/de6be627-d953-4024-9bcf-af39caf37f08). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 393554397
Change-Id: I9dcb01e8edbcca1c6ebafe2a7a5cc4684dcd153b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com