Revert "libvm_payload: Use Rust stubs instead of cc"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/71b389dd-b581-4954-bd39-ece79be6b096). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 393554360
Change-Id: Ib98fab715a6ee96983c584c5c820a05cdd5912d5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com