Revert "libvm_payload: Use Rust stubs instead of cc"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/de4f2b6e-2ecd-46d3-97c0-e9a845811967). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 393554373
Change-Id: I13bd16ebf512321eaf753a6bdb9a011126ceb4ae
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com