Revert "build/debian: Fix the resolution of build script's directory"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6843aae6-773f-4d6b-9973-de7cf5fbe799). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4e38d2e103db7215c747d4ac13ec08d5f7d76215
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com