virtmgr: ExtraDeviceTreeProperties
Creates an explicit type for the properties that are added to the guest
DT. This change makes the properties that are added to the guest DT more
explicit, as well as making it more obvious that *only* the secretkeeper
public key is fetched from the host DT.

On devices that support secretkeeper HAL v2, the dependency on the host
VM reference DT is removed entirely.

Note: no functional change intended.

Bug: 358078056
Test: TH
Change-Id: Id2f1b6aeb2cda5a481f0cf4a16481af3ef9dbcad