Revert "pvmfw: avb: Add API tests for tampered initrd"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b3399c1e-cfd7-4830-b90c-2de153b71b7b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iebb0b32361fa1a3ab69eec49ebc89248cbb17040
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com