Revert "libvm_payload: Use Rust stubs instead of cc"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3c62aaa0-1bf1-40c1-b3a8-74f26fb87bed). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2a6fe7299f0cdbfcd8481f015c80af078f78925e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com