Revert "Use new bssl_crypto inline static fn bindings"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3c62aaa0-1bf1-40c1-b3a8-74f26fb87bed). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic34713031e49d0f304db0e82a1a572a478c1cd47
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com