If rollback_index is specified use it in the dice chain of pVM
Ideally the payload node of the dice chain should specify both version
code of the apk and the rollback index. There even is a separate field
in the dice spec for that called security version code. However, this
field doesn't seem to be used for subcomponents of the payload node, so
for now we are overriding the version code with the value of the
rollback index.

Bug: 378681279
Test: atest MicrodroidTests
Change-Id: I3c4ecb0f7d725bd76347bdac1caa62b3c9106e06