Revert "Set different affinity for display activity"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/631a11b2-4a3c-4f8f-a348-c8e68b7b6ea5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I19a1cb39d1ebeb91b0e075f9bcef5044ce70e489
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com