Revert "Run weston as a systemd user service"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/631a11b2-4a3c-4f8f-a348-c8e68b7b6ea5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7d2b85050e172191f90d7025563a70c8c235cbef
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com