Revert "Clarify rollback protected API's documentation"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8e8fc1d3-6b3d-4d48-9ec4-3b7b4a05ab6b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8f6ecf9993688b5fe958401a5960257c0a35dde9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com