Change apex_available value of libavf based on avf_enabled
When avf_enabled is set libavf is included in com.android.virt,
but when it is unset libavf is not included in any apex.  That
triggered a shortcut in Soong when no apex variations exist
resulting in the platform variant not being hidden from Make
despite the module not setting //apex_available:platform.
When the shortcut is fixed to consistently hide the platform
variant it triggers a check-abi-dump-list failure because
libavf is now hidden from Make and not included in the LLNDK
libraries in LSDUMP_PATHS.

Add a conditional to libavf so that when avf_enabled it sets
apex_available: ["com.android.virt"] as before and uses the apex variant
to create the lsdump, but when avf_enabled is unset leaves apex_available
unset so that the platform variant exists and is unhidden to create the
lsdump.

Bug: 394664568
Test: builds
Change-Id: Id85edd1cb9548074537845137d66813e145642ea