Clean up emergency console and use panic for unexpected exceptions.
This removes `eprintln`, which was easy to misuse and not really sound.
Instead, introduce an unsafe function `emergency_console` with a clearly
documented safety requirement that it can only be called when the normal
UART instance will never be used again, e.g. just before rebooting.
Changed console_writeln! to return an error rather than panicking, so
panic handler can ignore the error.

Test: atest vmbase_example.integration_test
Change-Id: I76ac1911cf905fde0010054cfd8bc239699298f6