Revert "Change apex_available value of libavf based on avf_enabled"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fa918597-40ae-4c91-869f-2a3219b2228d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 394990888
Change-Id: I6cb3d6aef2923041a4822d053772d2cb73a05333
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com