Revert "Change apex_available value of libavf based on avf_enabled"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3dfd918e-99e9-4056-97b2-71cd580b86c2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 394999364
Change-Id: Ide8b7aa19f099362852787c7ff65ad295848b728
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com