Revert "Clarify rollback protected API's documentation"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/16d97d65-a596-49a6-bce7-4c93aad54968). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8be0ebe7a6189fd5e5c328bae333962405fbcbab
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com