Revert "Argument list for 0 arguments should be (void)"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/16d97d65-a596-49a6-bce7-4c93aad54968). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib9b424fa89d269d891fa118f933ca8ab657e3ed3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com