Revert "Clarify rollback protected API's documentation"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ed829500-a300-4d38-8f0b-334d10ecfcdf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If804dc06751a4e08c5af89869d4a4310a1b3ba56
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com