Revert "tests: pvmfw: export bcc.dat as a filegroup"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/39c367d9-f0f8-4790-a58f-6893b15bbd28). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I071e5f69027dcbe56e74f112f3f7805aaa311f00
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com