Revert "libavf: implement AVirtualMachineRawConfig_addCustomMemo..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dd923e4e-effe-4a8a-b72e-c4205b8b0a51). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 395679332
Change-Id: I160eb1cd9bfef508788b28d732f28d587c889151
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com