Revert "tests: pvmfw: export bcc.dat as a filegroup"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3b26e903-24c2-48f5-b2c0-36895729919f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0899bd6e3b57e197442d4d908fb253243985a76d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com