getFuncDurations to KvmHypTracer
When func and func_ret are enabled, allow to get a teardown of a
section. Those events will most likely generate a lot of traffic. So add
an option to set the function we don't want to trace, as well as a way
of setting a customized buffer.

Finally, let getDurationStats() work with any type of events mixed with
the required hyp_enter/hyp_exit.

Change-Id: I1b8cdeb4ca8d74285d2b68ec4fdfa9a2b6654f7a