pvmfw: Introduce Capability::TrustedWithRemoteKeys
Add a new capability allowing VM owners to instruct pvmfw to decorate
the DICE chain with the RKP VM marker.

As the marker instructs the RKP server to trust the guest payload, when
present throughout the DICE chain, pvmfw plays a crutial role in the
security of the released keys. Therefore, explicitly state our trust
assumptions by introducing a (currently redundant) extra check on the
signing authority of the guest payload, to prevent a future change from
inadvertently compromising the RKP server.

Use this capability for the RKP VM and the Trusty Security VM.

Bug: 391620545
Bug: 395794671
Test: m pvmfw
Test: atest libpvmfw_avb.integration_test libpvmfw.dice.test
Change-Id: I790579353324c3749650bfd385a80649683076fb