Revert "build/debian: Remove default kernel meta packages when d..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/97cb9ae5-e7e9-4197-b5d0-aacf3e5e436c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4dd764803bcd25756d154e4c7ebcc7d4dd775b94
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com