Revert "tests: pvmfw: export bcc.dat as a filegroup"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0a492319-b50d-44b2-bf88-cf3b1cb771b4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6684b7ef048d5f6b3aa686fb08d0fdda089cbf36
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com