Revert "libavf: implement AVirtualMachineRawConfig_addCustomMemo..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3baf77ed-2821-421a-98d9-5d073939af03). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 396079672
Change-Id: Id3eec52331d7594bf9e103228b9f75757a3b4165
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com