Revert "apkmanifest: parse fields related to relaxed rollback pr..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8f3f801e-bac8-42fc-9d02-9593e645f28d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I78ded3896c6466c011bcd9bc2ed4832602c5185e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com