Revert "tests: pvmfw: export bcc.dat as a filegroup"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f77e490b-86f3-48c8-844d-abb37eb5d3e6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib2ff327fb7f63e81e5c94e2974c3cd46cf829617
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com