Revert "virtmgr: fix /system/[system_ext/product] path rebase"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/46f53939-0f4a-4b74-b58f-cedaa6603ea7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic50a8978439f7c4321a19bc9513392bac8b5197a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com