Revert "virtmgr: fix /system/[system_ext/product] path rebase"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8ecfac8d-fcfd-4888-bb22-fbffe8dcfd7b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5490a568e64cc2cf2149b49af5e10028f61f3be9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com