Revert "virtmgr: loosen kernel/initrd SELinux requirements for v..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5cae6caf-50e6-48b6-815c-2bfbe3bbd1d5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I88f69b0af3e0e1d1e0778a1b4e6bfd4baf0a98df
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com