Revert "tests: pvmfw: export bcc.dat as a filegroup"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5d92e0a4-792b-49a3-a310-2ac2d903e66e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I54f05764013af6bf2883fbadf60f39fd8f5f061c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com