Revert "Move requestPermissions from onResume to onCreate"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fa5c874f-a63f-4b81-9eae-3bcff39d2680). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I382696309e18da9929f87c6d2b516f79a3e8a457
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com