Implement memory ballooning in Terminal App
This change implements memory ballooning in the Terminal App to
dynamically adjust the memory allocated to the VM.

The Application class now includes an ApplicationLifecycleObserver which
binds to the VmLauncherService and sends lifecycle events (onStart,
onStop). The VmLauncherService then calls setMemoryBalloonByPercent() on
the VirtualMachine instance to adjust the memory. On application start
(onStart), the memory balloon is deflated (0%), maximizing available
memory for the app. On application stop (onStop), the memory balloon is
inflated (50%), allowing the system to reclaim memory while the app is
in the background.

A new API, memoryBalloonEnabled(), has been added to
IVirtualMachine.aidl to check if the balloon is enabled before
attempting to set its size, and a corresponding implementation has been
added to VirtualMachine.java.

Bug: b/392791968
Test: Maunual test memory balloon works in Terminal App
Test: Verify setMemoryBalloon() is not called if memory balloon disabled

Change-Id: I26e9bd944b29c27831f54be8b4ddba8c29020a5a