Revert "Add warning message when AVF RKP HAL is not declared"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a4817f1e-0123-4757-8fee-ea9f994b1001). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I03ac448ad185e9cdaf063688a5bb578423646da1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com