AVF: allow FD -> IBinder.
Okay, so before we have two APIs:
- connectToVsockServer "AB"
- connectVsock "A"

This adds:
- binderFromPreconnectedClient "B"

That is:
- connectToVsockServer = connectVsock + binderFromPreconnectedClient

Normally, in the same process, you do all of this, but to
pass a connection to another process, imagine
processes "Q -> R -> S", where:
- Q is client
- R is VM creator
- S is VM

Then Q can call binderFromPreconnectedClient, and it can
implement the provider by calling an API on R which will
return the results of R calling connectVsock.

Non-binder use of connectVsock is not recommended.

Bugs: me and without this, extra hops cost performannce
Test: atest MicrodroidTestApp - several tests pass, so letting TH run
the rest

Change-Id: I670243714bbb6167a1fdae2bb898b1f5a9d01f5f