Revert "Remove debugging output"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/85c11ae9-2704-4eba-969e-35175209f805). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iba2f1d1cdae740d322115089787e646e5f6fbff4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com