Revert "Logs from the terminal VM is accumulated"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/46a70879-5cd6-4270-8a50-295fb618b72f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 401685022
Change-Id: I0785e99137f198431d15db470a4ca67d4ccb455a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com