Revert "[trusty] Add trusty_vm helper script targets for reuse"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d5f9bdc1-9565-44a6-8107-e6d0bc9ced85). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 402143900
Change-Id: I2501535a2aefb94e68b6b12c987cbe82041d4a10
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com