Revert "[trusty] Add trusty_vm helper script targets for reuse"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/537f7252-e5d1-4bb6-924c-fa1e6ab38f60). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 402144142
Change-Id: Ib5864b6bfe9fb852dc321e3f7dcf0940ea7718fa
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com