Revert "[trusty] Add trusty_vm helper script targets for reuse"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c74a3fd8-bcd9-473a-8f60-e3115b6d52f6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 402144281
Change-Id: I4a1f4a6b495b794e6ab5eb154447c87b62229edc
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com