Revert "Retrieve terminal info in a more robust way"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c7fb9a42-4dbe-42b9-86dc-f0d146885022). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id35a6a49ff0b4fea923926e11f787e879b80d948
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com