Revert "Timeout for VM start"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5ea15a51-a09e-4b34-a0f2-ef97595daf79). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 402727622
Change-Id: I611d99b4a9f85f813476d02540fccbafd292d20a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com