Revert "Timeout for VM start"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e544cc0a-357c-4443-983e-a04a20628cb5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 402873093
Bug: 402875693
Change-Id: I2c6c88d0d08442c559092e7eeabd08b6d1247610
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com