Revert "virtmgr: add CallingPartition enum"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fdafb3ba-c355-4e31-9bf3-68343670ff6b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0fd02a955d75780b9789475809f7cd0b31f36e44
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com