Revert "virtmgr: loosen kernel/initrd SELinux requirements for v..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fdafb3ba-c355-4e31-9bf3-68343670ff6b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If20d631b42ce22f03bc5b86904c154a08f66912a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com