Revert "virtmgr: fix /system/[system_ext/product] path rebase"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/39e7ccf4-11e4-4160-a697-7dcc4ab5e882). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4c57fbfab824fcd57e26b20858ca5dc0866354a6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com