Revert "[dice] Accept different hash sizes for VM DICE chain"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a4f4cbc6-ac9e-4141-a735-46209c167920). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2ee1009dcb216488fe7789fbd08cbad08a575a8f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com