Allow override of the trusty system vm packages
This will be overridden by desktop emulator configs to switch out the
trusty vm thas is loaded.

Due to the nature of lazy expansion in Make, the `+=` does *not*
immediately expand the `?=`. This can be observed by running `make test`
with the following Makefile:

```
p = a b
x ?= x y z
p += $(x)
x := c d

test:
	@echo $(p)
```

Bug: 398934069
Test: launch_cvd unaffected
Change-Id: I631b6da2e8db2b4b5c66ecff37d0bfedf07fc1aa