Revert "Focus to TerminalView when switching tabs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5ddc6a9d-a4a3-4722-8413-f27885e39d73). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I69f075d762e1346e20ea7847be115d4568baac45
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com