Revert "[dice] Accept different hash sizes for VM DICE chain"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9b925135-3926-4757-a4d1-1f723d74c1ea). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3a9ae8ae6d8eaba993309b0e0be672d9c0a81cea
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com