Revert "Focus to TerminalView when switching tabs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3b61881d-27f4-475f-be32-ccd05f1cadfc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4c14e85afc0e2375094f5a3fc6079c10412c7d7a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com