Revert "VirtualMachine: Clean up input event executor when disco..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0d16a780-f48a-4a6a-b35b-35ee7d2605dc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7bf72a923e6ab49e38dfbd6d852c16a1656c5245
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com