Revert "guest: trusty: test_vm_os: add a dedicated VM for testin..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3e5a1999-759f-4f1a-b6f9-f9f55f411934). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia10fabe43a711fe700341448ec8c17b4fb009322
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com