Revert "Reduce stopSelf and ACTION_SHUTDOWN_VM"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fe56622a-c1c7-47b1-911a-2ce19499430e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9a24794784579ba5c153ebbb9f932ec2f85e27ab
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com