Revert "Add trusty dirgroup to sample_inputs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/87729b68-492a-4289-a322-3fc5f5c92b1d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I855b140b9897477032281fb12279b643ce09aec8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com