Revert "Focus to TerminalView when switching tabs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3e93273f-dcca-4417-993d-5ab962c71a61). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If804a031a6acab5f1bbc1e0a9d6e8f8cdfc76a1b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com