Revert "Remove legacy debian.log file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/141c478b-8b06-4d2a-b723-5281b8a32ea0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I63602887eb3cd74d27ffdd18f2c7ffd7e8748935
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com