Revert "TerminalApp: Remove onDestroy() from ApplicationLifecycl..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/43da5687-eeea-40f1-8421-417e2788d75c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1867e7857e61824f81636628a91bb54e949a3f76
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com