Revert "build/debian: Allow booting the kernel directly"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fe8069a0-cc93-47a1-99eb-2f60b2adbf76). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iefac8f37ad0e94bc06f060c0a4641ce58395e09a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com