Revert "Use safe-mmio crate for 8250 UART driver."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a078e371-3859-4417-a8d9-1e349396609b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib107237713c6187c00e39d6d92160703cf25cb57
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com