Revert "Use safe-mmio crate for 8250 UART driver."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1f75818d-9b7f-4710-bdd4-22ce48396688). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3a65d38e88b945947610188064d85a5b48057a01
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com