Revert "Use safe-mmio crate for 8250 UART driver."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/25a667ae-eb23-41cd-8285-8d2b49a15000). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib61d262047ab6d020a19fd6287caeb483cd05b9c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com