Revert "Focus to TerminalView when switching tabs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ea62351e-c78e-435b-9497-5ecf0f938318). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2acb21bbebcc09464b7eb8e6a6057d0f41aa813d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com