Revert "Use safe-mmio crate for 8250 UART driver."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/07df6fb0-501b-4ce5-8577-59607dae606d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8a1d8ed65d5bf63fe3df9f53e3f3271322835529
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com