Revert "Use safe-mmio crate for 8250 UART driver."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9a320c92-c5bb-4721-822c-589365972d3f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibfd3d0a61d8317fb4ddb8f8115d914e075462451
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com