Revert "Terminal: Enable zram"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e8ecc06f-1fef-41a0-8fe3-21a3f11ce7f5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2a96e1763a92434436ed426e6bc3aaabac813906
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com