Revert "TerminalApp: Remove onDestroy() from ApplicationLifecycl..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b17bc940-f41f-468c-8dc4-8f6d13da2e7e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I802c8daf9cf7c35784ac6a0e5315ceb469c23cc7
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com