Revert "Remove the cap of max disk resize."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3788e659-c449-437e-83da-e9258a147411). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id543a36faedfd2de61bcf87963b9a36ff4d428a1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com