Revert "guest: debian: Implement storage_balloon_agent"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d9ccc801-f60b-4261-8292-f14d7c2657ed). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9b647590646c058eec8dc83f68ef7969f2990e49
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com