Revert "Terminal: Enable zram"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/04aaa546-e074-43cb-a6c6-c9bf12a05e38). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I11207ff5bf7d51dd648d681616d8e90b9c8ee1ff
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com