Revert "Focus to TerminalView when switching tabs"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4c0a3b49-61d1-423d-84f9-3163ee95330d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I83914e88e898bc4cd388627c8220777d02b828e8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com