Revert "pvmfw: README: Add definition for "pVM""
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4b022a1b-5487-4aee-b761-8bed239fc9e6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I05774e06b7c467a4efb023755cd5c65537c5a580
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com