Don't run supportForProtectedOrNonProtectedVms on GSI
Background:

The test ensures that either protected VMs or non-protected VMs are
supported "IF" AVF is enabled on the device. In other words, the test is
skipped on devices where AVF is not enabled (because their BSP is old).

Problem:

When such a device runs CTS-on-GSI, the system image is replaced with
GSI and AVF is "unconditionally" enabled there. Therefore this test is
not skipped, and fails.

Fix:

Skip this test on GSI.

Bug: 411050619
Test: N/A
Flag: EXEMPT test bug fix
Change-Id: I653e81a1466391c9d962e99abc96b48a0e27dc00