Revert "Update a test expectation."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8563f45f-45ce-4a05-8af7-1f79367b5872). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If3ba2caee007ac3d02b84ab0b49a545f3772379d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com