Revert "Clarify rollback protected API's documentation"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bb22c522-4d55-4a03-afba-b7891f11338e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0541e5227f207eaf07ea3b6c8ce3836146c95259
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com